+++ src/routes/v1/GetProfileWithStrategy.ts
@@ -1,12 +1,12 @@
 import { Request, Response } from '../../types/express'
-import { SkyBlockProfileStats } from '../../types/hypixel'
-import { asyncWrap, parseSkyBlockProfiles, preformGetRequest, validateUuid } from '../../utils'
+import { PlayerStats, SkyBlockProfileStats } from '../../types/hypixel'
+import { asyncWrap, mergeSkyBlockProfileAndPlayer, parseHypixelPlayer, parseSkyBlockProfiles, preformGetRequest, validateUuid } from '../../utils'
 
 export default asyncWrap(async (request: Request, response: Response) => {
   const uuid = validateUuid(request.params.uuid)
 
-  const hypixelResponse = await preformGetRequest(response, `/skyblock/profiles?key=${request.authToken}&uuid=${uuid}`)
-  const profiles: SkyBlockProfileStats[] = parseSkyBlockProfiles(hypixelResponse, uuid)
+  const profilesResponse = await preformGetRequest(response, `/skyblock/profiles?key=${request.authToken}&uuid=${uuid}`)
+  const profiles: SkyBlockProfileStats[] = parseSkyBlockProfiles(profilesResponse, uuid)
 
   let profile: SkyBlockProfileStats | undefined
   switch (request.params.strategy) {
@@ -67,9 +67,12 @@ export default asyncWrap(async (request: Request, response: Response) => {
     })
   }
 
+  const playerResponse = await preformGetRequest(response, `/player?key=${request.authToken}&uuid=${uuid}`)
+  const player: PlayerStats = parseHypixelPlayer(playerResponse, uuid)
+
   return response.status(200).json({
     status: 200,
-    data: profile,
+    data: mergeSkyBlockProfileAndPlayer(profile, player),
   })
 })