+++ src/types/hypixel/PlayerStats.d.ts
@@ -1,0 +1,12 @@
+export default interface PlayerStats {
+  username: string
+  firstLogin: number
+  lastLogin: number
+  socialMedia: {
+    links?: {
+      DISCORD?: string
+      TWITTER?: string
+      TWITCH?: string
+    }
+  }
+}