+++ src/utils/index.ts
@@ -8,4 +8,6 @@ export { humanizeTime as humanizeTime } from './Time'
 export { isUuid as isUuid } from './Uuid'
 export { validateUuid as validateUuid } from './Uuid'
 
-export { parseSkyBlockProfiles as parseSkyBlockProfiles } from './SkyBlock'
+export { mergeSkyBlockProfileAndPlayer as mergeSkyBlockProfileAndPlayer } from './Hypixel'
+export { parseHypixelPlayer as parseHypixelPlayer } from './Hypixel'
+export { parseSkyBlockProfiles as parseSkyBlockProfiles } from './Hypixel'